# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Book Catalogue API
  description: >
    REST API for managing a library catalogue: create, read, update and delete
    books with validation, pagination, caching and uniform error handling.
  version: v1
  contact:
    name: Enterprise Web Application Development - Kabul University
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
tags:
  - name: Books
    description: CRUD operations for the library catalogue
paths:
  /api/v1/books:
    get:
      tags: [Books]
      summary: List all books
      description: Returns a paginated and sortable page of the catalogue.
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 0, minimum: 0 }
          description: Zero-based page index.
        - name: size
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
          description: Page size (capped by spring.data.web.pageable.max-page-size).
        - name: sort
          in: query
          schema: { type: string, example: "title,asc" }
          description: Sort property and direction; sortable fields are id, title,
            author, isbn, publishedYear, category, createdAt, updatedAt.
      responses:
        "200":
          description: Page of books
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PageResponseBookResponse" }
        "400":
          description: Invalid pagination or sort parameter
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    post:
      tags: [Books]
      summary: Create a new book
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookRequest" }
            example:
              title: Refactoring
              author: Martin Fowler
              isbn: "9780134757599"
              publishedYear: 2018
              category: Programming
      responses:
        "201":
          description: Book created (Location header points at the new resource)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookResponse" }
        "400":
          description: Validation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: ISBN already exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api/v1/books/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer, format: int64 }
        description: Book identifier.
    get:
      tags: [Books]
      summary: Get a book by id
      responses:
        "200":
          description: The book
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookResponse" }
        "404":
          description: Book not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    put:
      tags: [Books]
      summary: Update an existing book
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookRequest" }
      responses:
        "200":
          description: Book updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookResponse" }
        "400":
          description: Validation failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "404":
          description: Book not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
        "409":
          description: ISBN already in use by another book
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      tags: [Books]
      summary: Delete a book
      responses:
        "204":
          description: Book deleted
        "404":
          description: Book not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
components:
  schemas:
    BookRequest:
      type: object
      required: [title, author, isbn, publishedYear, category]
      properties:
        title: { type: string, maxLength: 200 }
        author: { type: string, maxLength: 100 }
        isbn:
          type: string
          description: 13-digit ISBN-13 (hyphens/spaces tolerated, check digit verified)
          example: "9780134757599"
        publishedYear: { type: integer, minimum: 1900, example: 2018 }
        category: { type: string, maxLength: 100, example: Programming }
    BookResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        title: { type: string }
        author: { type: string }
        isbn: { type: string, example: "9780134757599" }
        publishedYear: { type: integer }
        category: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PageResponseBookResponse:
      type: object
      properties:
        content:
          type: array
          items: { $ref: "#/components/schemas/BookResponse" }
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        first: { type: boolean }
        last: { type: boolean }
        hasNext: { type: boolean }
        hasPrevious: { type: boolean }
        sort: { type: string, example: "id,asc" }
    ErrorResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer, example: 404 }
        error: { type: string, example: Not Found }
        message: { type: string }
        path: { type: string }
        validationErrors:
          type: object
          additionalProperties: { type: string }
          description: Present only for bean-validation failures (field -> message).